refactor(broker): extract creation of raft partition group
