+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/PartitionManagerImpl.java
@@ -16,7 +16,6 @@ import io.atomix.primitive.partition.ManagedPartitionService;
 import io.atomix.primitive.partition.impl.DefaultPartitionService;
 import io.atomix.raft.partition.RaftPartition;
 import io.atomix.raft.partition.RaftPartitionGroup;
-import io.atomix.raft.partition.RaftPartitionGroup.Builder;
 import io.atomix.utils.concurrent.Futures;
 import io.camunda.zeebe.broker.PartitionListener;
 import io.camunda.zeebe.broker.clustering.ClusterServices;
@@ -29,9 +28,6 @@ import io.camunda.zeebe.broker.partitioning.topology.TopologyManagerImpl;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyPartitionListener;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyPartitionListenerImpl;
 import io.camunda.zeebe.broker.system.configuration.BrokerCfg;
-import io.camunda.zeebe.broker.system.configuration.ClusterCfg;
-import io.camunda.zeebe.broker.system.configuration.DataCfg;
-import io.camunda.zeebe.broker.system.configuration.NetworkCfg;
 import io.camunda.zeebe.broker.system.management.deployment.PushDeploymentRequestHandler;
 import io.camunda.zeebe.broker.system.monitoring.BrokerHealthCheckService;
 import io.camunda.zeebe.broker.system.monitoring.DiskSpaceUsageListener;
@@ -63,18 +59,12 @@ import io.camunda.zeebe.engine.processing.streamprocessor.ProcessingContext;
 import io.camunda.zeebe.engine.processing.streamprocessor.TypedRecord;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.CommandResponseWriter;
 import io.camunda.zeebe.engine.state.mutable.MutableZeebeState;
-import io.camunda.zeebe.logstreams.impl.log.ZeebeEntryValidator;
 import io.camunda.zeebe.logstreams.log.LogStream;
 import io.camunda.zeebe.protocol.impl.encoding.BrokerInfo;
-import io.camunda.zeebe.snapshots.ReceivableSnapshotStoreFactory;
 import io.camunda.zeebe.snapshots.impl.FileBasedSnapshotStoreFactory;
-import io.camunda.zeebe.util.FileUtil;
 import io.camunda.zeebe.util.health.HealthStatus;
 import io.camunda.zeebe.util.sched.ActorControl;
 import io.camunda.zeebe.util.sched.ActorSchedulingService;
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
@@ -165,7 +155,8 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
     this.commandHApiService = commandHApiService;
     this.exporterRepository = exporterRepository;
 
-    partitionGroup = buildRaftPartitionGroup(brokerCfg, snapshotStoreFactory);
+    partitionGroup =
+        new RaftPartitionGroupFactory().buildRaftPartitionGroup(brokerCfg, snapshotStoreFactory);
 
     final var membershipService = clusterServices.getMembershipService();
     final var communicationService = clusterServices.getCommunicationService();
@@ -226,71 +217,6 @@ public final class PartitionManagerImpl implements PartitionManager, TopologyMan
           new ZeebePartition(transitionContext, transitionBehavior);
       partitions.add(zeebePartition);
     }
-  }
-
-  private static RaftPartitionGroup buildRaftPartitionGroup(
-      final BrokerCfg configuration, final ReceivableSnapshotStoreFactory snapshotStoreFactory) {
-
-    final DataCfg dataConfiguration = configuration.getData();
-    final String rootDirectory = dataConfiguration.getDirectory();
-    final var rootPath = Paths.get(rootDirectory);
-    try {
-      FileUtil.ensureDirectoryExists(rootPath);
-    } catch (final IOException e) {
-      throw new UncheckedIOException("Failed to create data directory", e);
-    }
-
-    final var raftDataDirectory = rootPath.resolve(GROUP_NAME);
-
-    try {
-      FileUtil.ensureDirectoryExists(raftDataDirectory);
-    } catch (final IOException e) {
-      throw new UncheckedIOException("Failed to create Raft data directory", e);
-    }
-
-    final ClusterCfg clusterCfg = configuration.getCluster();
-    final var experimentalCfg = configuration.getExperimental();
-    final DataCfg dataCfg = configuration.getData();
-    final NetworkCfg networkCfg = configuration.getNetwork();
-
-    final Builder partitionGroupBuilder =
-        RaftPartitionGroup.builder(GROUP_NAME)
-            .withNumPartitions(clusterCfg.getPartitionsCount())
-            .withPartitionSize(clusterCfg.getReplicationFactor())
-            .withMembers(getRaftGroupMembers(clusterCfg))
-            .withDataDirectory(raftDataDirectory.toFile())
-            .withSnapshotStoreFactory(snapshotStoreFactory)
-            .withMaxAppendBatchSize((int) experimentalCfg.getMaxAppendBatchSizeInBytes())
-            .withMaxAppendsPerFollower(experimentalCfg.getMaxAppendsPerFollower())
-            .withEntryValidator(new ZeebeEntryValidator())
-            .withFlushExplicitly(!experimentalCfg.isDisableExplicitRaftFlush())
-            .withFreeDiskSpace(dataCfg.getFreeDiskSpaceReplicationWatermark())
-            .withJournalIndexDensity(dataCfg.getLogIndexDensity())
-            .withPriorityElection(experimentalCfg.isEnablePriorityElection());
-
-    final int maxMessageSize = (int) networkCfg.getMaxMessageSizeInBytes();
-
-    final var segmentSize = dataCfg.getLogSegmentSizeInBytes();
-    if (segmentSize < maxMessageSize) {
-      throw new IllegalArgumentException(
-          String.format(
-              "Expected the raft segment size greater than the max message size of %s, but was %s.",
-              maxMessageSize, segmentSize));
-    }
-
-    partitionGroupBuilder.withSegmentSize(segmentSize);
-
-    return partitionGroupBuilder.build();
-  }
-
-  private static List<String> getRaftGroupMembers(final ClusterCfg clusterCfg) {
-    final int clusterSize = clusterCfg.getClusterSize();
-    // node ids are always 0 to clusterSize - 1
-    final List<String> members = new ArrayList<>();
-    for (int i = 0; i < clusterSize; i++) {
-      members.add(Integer.toString(i));
-    }
-    return members;
   }
 
   private PartitionBoostrapAndTransitionContextImpl buildTransitionContext(
